convert mtree to use new job pool
NB: valgrind is picking up a problem at job.c:53 of job_push() which I'm
not sure about.  It seems to be claiming that the job is free()'d by the
callback in job_seek() *before* it has its pointer set in job_push()
which seems unlikely.  Threads sometimes confuse valgrind, so it could
be a false positive.  All the tests pass regardless, which suggests the
hashing is all happening in the correct order.  Further investigation
needed and perhaps a suppression adding unless a bug can be found.